MNT: Remove unused code paths (#1965)
* TST: pandas==1.0 compat

* Remove old matplotlib paths

* Remove old scipy paths

* Remove _set_spine_position

The behavior that this function addressed has been resolved in matplotlib 1.4:
https://github.com/matplotlib/matplotlib/pull/3104

* Clean more old compat code paths

Removed a few try/except clauses aimed to cover old matplotlib/pandas
versions that are no longer supported.

* Remove __future__ imports

* Improve code style

+++ seaborn/tests/test_relational.py
@@ -1,4 +1,3 @@
-from __future__ import division
 from itertools import product
 import warnings
 import numpy as np
@@ -122,7 +121,7 @@ class TestRelationalPlotter(object):
     @pytest.fixture
     def null_column(self):
 
-        return pd.Series(index=np.arange(20))
+        return pd.Series(index=np.arange(20), dtype='float64')
 
     def test_wide_df_variables(self, wide_df):
